# EagleBank/wwwroot/swagger/v1/swagger.yaml
openapi: 3.0.4
info:
  title: 'EagleBank, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null'
  version: '1.0'
paths:
...
  /v1/users:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: OK
  /v1/users/login:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UserDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserDto'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LoginDto'
            application/json:
              schema:
                $ref: '#/components/schemas/LoginDto'
            text/json:
              schema:
                $ref: '#/components/schemas/LoginDto'
  '/v1/users/{id}':
    get:
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
            text/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
components:
  schemas:
...
    LoginDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        token:
          type: string
          nullable: true
      additionalProperties: false
...
    UserDto:
      required:
        - password
        - username
      type: object
      properties:
        username:
          minLength: 1
          type: string
        password:
          minLength: 1
          type: string
      additionalProperties: false
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
          nullable: true
      additionalProperties: false